Relocate some middleware
